refactor(metadata): address CodeRabbit review findings
- Add keyboard accessibility (Enter/Space) to duplicate-indicator button
- Extract string similarity utilities to reusable module
- Fix timer leak by tracking and clearing duplicate check timeout
- Fix matchWindow calculation using Math.floor for correct integer bounds
- Update tests to import utilities directly instead of duplicating logic

All 33 tests pass with updated similarity scores.
