+++ frontend/src/app/features/metadata/component/metadata-manager/metadata-manager.component.spec.ts
@@ -1,98 +1,5 @@
 import {describe, expect, it} from 'vitest';
-
-// Test the duplicate detection algorithms directly
-// These are standalone test implementations of the algorithms
-
-// Test implementations of the algorithms
-const calculateJaro = (s1: string, s2: string): number => {
-  if (s1 === s2) return 1.0;
-  if (s1.length === 0 || s2.length === 0) return 0.0;
-
-  const matchWindow = Math.max(s1.length, s2.length) / 2 - 1;
-  const s1Matches = new Array(s1.length).fill(false);
-  const s2Matches = new Array(s2.length).fill(false);
-
-  let matches = 0;
-  let transpositions = 0;
-
-  // Find matches
-  for (let i = 0; i < s1.length; i++) {
-    const start = Math.max(0, i - matchWindow);
-    const end = Math.min(i + matchWindow + 1, s2.length);
-
-    for (let j = start; j < end; j++) {
-      if (s2Matches[j] || s1[i] !== s2[j]) continue;
-      s1Matches[i] = true;
-      s2Matches[j] = true;
-      matches++;
-      break;
-    }
-  }
-
-  if (matches === 0) return 0.0;
-
-  // Find transpositions
-  let k = 0;
-  for (let i = 0; i < s1.length; i++) {
-    if (!s1Matches[i]) continue;
-    while (!s2Matches[k]) k++;
-    if (s1[i] !== s2[k]) transpositions++;
-    k++;
-  }
-
-  return (matches / s1.length + matches / s2.length + (matches - transpositions / 2) / matches) / 3.0;
-};
-
-const calculateJaroWinkler = (s1: string, s2: string): number => {
-  const jaro = calculateJaro(s1, s2);
-
-  if (jaro < 0.7) {
-    return jaro;
-  }
-
-  let prefixLength = 0;
-  const maxPrefix = Math.min(4, s1.length, s2.length);
-  for (let i = 0; i < maxPrefix; i++) {
-    if (s1[i] === s2[i]) {
-      prefixLength++;
-    } else {
-      break;
-    }
-  }
-
-  return jaro + (prefixLength * 0.1 * (1 - jaro));
-};
-
-const areAuthorsSimilar = (name1: string, name2: string, threshold: number): boolean => {
-  const lower1 = name1.toLowerCase();
-  const lower2 = name2.toLowerCase();
-
-  const parts1 = lower1.split(/\s+/).filter(p => p.length > 0);
-  const parts2 = lower2.split(/\s+/).filter(p => p.length > 0);
-
-  if (parts1.length === 1 && parts2.length === 1) {
-    return calculateJaroWinkler(lower1, lower2) >= threshold;
-  }
-
-  if (parts1.length === 1 || parts2.length === 1) {
-    return false;
-  }
-
-  const firstName1 = parts1[0];
-  const firstName2 = parts2[0];
-  const lastName1 = parts1[parts1.length - 1];
-  const lastName2 = parts2[parts2.length - 1];
-
-  const firstNameSimilarity = calculateJaroWinkler(firstName1, firstName2);
-  const lastNameSimilarity = calculateJaroWinkler(lastName1, lastName2);
-
-  const MIN_PART_THRESHOLD = 0.75;
-  const averageSimilarity = (firstNameSimilarity + lastNameSimilarity) / 2;
-
-  return firstNameSimilarity >= MIN_PART_THRESHOLD &&
-         lastNameSimilarity >= MIN_PART_THRESHOLD &&
-         averageSimilarity >= threshold;
-};
+import {calculateJaro, calculateJaroWinkler, areAuthorsSimilar} from '../../../../shared/utils/string-similarity.util';
 
 describe('MetadataManagerComponent - Duplicate Detection', () => {
 
@@ -109,8 +16,8 @@ describe('MetadataManagerComponent - Duplicate Detection', () => {
 
     it('should return moderate similarity for strings with typos', () => {
       const result = calculateJaroWinkler('stephen', 'stefen');
-      expect(result).toBeGreaterThan(0.6);
-      expect(result).toBeLessThan(0.7);
+      expect(result).toBeGreaterThan(0.89);
+      expect(result).toBeLessThan(0.90);
     });
 
     it('should apply Winkler bonus for matching prefixes', () => {
