+++ frontend/src/app/features/metadata/component/metadata-manager/metadata-manager.component.ts
@@ -18,6 +18,7 @@ import {ActivatedRoute, Router} from '@angular/router';
 import {Subscription} from 'rxjs';
 import {ExternalDocLinkComponent} from '../../../../shared/components/external-doc-link/external-doc-link.component';
 import {TranslocoDirective, TranslocoPipe, TranslocoService} from '@jsverse/transloco';
+import {areAuthorsSimilar} from '../../../../shared/utils/string-similarity.util';
 
 interface MetadataItem {
   value: string;
@@ -75,6 +76,7 @@ export class MetadataManagerComponent implements OnInit, OnDestroy {
   private readonly t = inject(TranslocoService);
 
   private routeSub!: Subscription;
+  private duplicateCheckTimer: ReturnType<typeof setTimeout> | null = null;
   private readonly syncMetadataEffect = effect(() => {
     const isLoading = this.bookService.isBooksLoading();
     this.loading.set(isLoading);
@@ -169,6 +171,10 @@ export class MetadataManagerComponent implements OnInit, OnDestroy {
 
   ngOnDestroy(): void {
     this.routeSub.unsubscribe();
+    if (this.duplicateCheckTimer !== null) {
+      clearTimeout(this.duplicateCheckTimer);
+      this.duplicateCheckTimer = null;
+    }
   }
 
   private extractMetadata(books: Book[]) {
@@ -207,10 +213,15 @@ export class MetadataManagerComponent implements OnInit, OnDestroy {
     this.languages = this.mapToItems(languagesMap);
 
     // Compute duplicates for authors (and optionally series)
-    setTimeout(() => {
+    // Clear any existing timer before scheduling a new one
+    if (this.duplicateCheckTimer !== null) {
+      clearTimeout(this.duplicateCheckTimer);
+    }
+    this.duplicateCheckTimer = setTimeout(() => {
       this.enrichWithDuplicates('authors');
       // Uncomment if you want duplicate detection for series as well:
       // this.enrichWithDuplicates('series');
+      this.duplicateCheckTimer = null;
     }, 0);
   }
 
@@ -659,9 +670,7 @@ export class MetadataManagerComponent implements OnInit, OnDestroy {
       for (let j = 0; j < items.length; j++) {
         if (i === j) continue;
 
-        const isSimilar = type === 'authors'
-          ? this.areAuthorsSimilar(item.value, items[j].value, THRESHOLD)
-          : this.calculateJaroWinkler(item.value.toLowerCase(), items[j].value.toLowerCase()) >= THRESHOLD;
+        const isSimilar = areAuthorsSimilar(item.value, items[j].value, THRESHOLD);
 
         if (isSimilar) {
           duplicates.push(items[j].value);
@@ -673,121 +682,4 @@ export class MetadataManagerComponent implements OnInit, OnDestroy {
     }
   }
 
-  private areAuthorsSimilar(name1: string, name2: string, threshold: number): boolean {
-    const lower1 = name1.toLowerCase();
-    const lower2 = name2.toLowerCase();
-
-    // Parse names into parts (first, middle, last)
-    const parts1 = lower1.split(/\s+/).filter(p => p.length > 0);
-    const parts2 = lower2.split(/\s+/).filter(p => p.length > 0);
-
-    // If both names are single words, use simple Jaro-Winkler
-    if (parts1.length === 1 && parts2.length === 1) {
-      return this.calculateJaroWinkler(lower1, lower2) >= threshold;
-    }
-
-    // If one is single-word and the other is multi-part, they should not match
-    // This prevents "Molière" from matching "Molière Hubert Carrier"
-    // (different people despite sharing a name component)
-    if (parts1.length === 1 || parts2.length === 1) {
-      return false;
-    }
-
-    // For multi-part names, check if both first AND last names are similar
-    // This prevents "Alessandro Baricco" from matching "Alessandro Volta"
-    const firstName1 = parts1[0];
-    const firstName2 = parts2[0];
-    const lastName1 = parts1[parts1.length - 1];
-    const lastName2 = parts2[parts2.length - 1];
-
-    const firstNameSimilarity = this.calculateJaroWinkler(firstName1, firstName2);
-    const lastNameSimilarity = this.calculateJaroWinkler(lastName1, lastName2);
-
-    // Both first and last names must meet a minimum threshold
-    // First name needs to be at least 0.75 similar, last name at least 0.75 similar
-    // And the average should meet the main threshold
-    const MIN_PART_THRESHOLD = 0.75;
-    const averageSimilarity = (firstNameSimilarity + lastNameSimilarity) / 2;
-
-    return firstNameSimilarity >= MIN_PART_THRESHOLD &&
-           lastNameSimilarity >= MIN_PART_THRESHOLD &&
-           averageSimilarity >= threshold;
-  }
-
-  private calculateJaroWinkler(s1: string, s2: string): number {
-    const jaro = this.calculateJaro(s1, s2);
-
-    // If the Jaro similarity is below a threshold, don't apply Winkler bonus
-    if (jaro < 0.7) {
-      return jaro;
-    }
-
-    // Calculate common prefix length (up to 4 characters)
-    let prefixLength = 0;
-    const maxPrefix = Math.min(4, s1.length, s2.length);
-    for (let i = 0; i < maxPrefix; i++) {
-      if (s1[i] === s2[i]) {
-        prefixLength++;
-      } else {
-        break;
-      }
-    }
-
-    // Jaro-Winkler formula: jw = jaro + (prefixLength * 0.1 * (1 - jaro))
-    return jaro + (prefixLength * 0.1 * (1 - jaro));
-  }
-
-  private calculateJaro(s1: string, s2: string): number {
-    if (s1 === s2) return 1.0;
-    if (s1.length === 0 || s2.length === 0) return 0.0;
-
-    const matchWindow = Math.max(s1.length, s2.length) / 2 - 1;
-    const { matches, s1Matches, s2Matches } = this.findMatches(s1, s2, matchWindow);
-
-    if (matches === 0) return 0.0;
-
-    const transpositions = this.countTranspositions(s1, s2, s1Matches, s2Matches);
-
-    // Jaro formula: (m/|s1| + m/|s2| + (m - t/2)/m) / 3
-    return (matches / s1.length + matches / s2.length + (matches - transpositions / 2) / matches) / 3.0;
-  }
-
-  private findMatches(s1: string, s2: string, matchWindow: number): {
-    matches: number;
-    s1Matches: boolean[];
-    s2Matches: boolean[];
-  } {
-    const s1Matches = new Array(s1.length).fill(false);
-    const s2Matches = new Array(s2.length).fill(false);
-    let matches = 0;
-
-    for (let i = 0; i < s1.length; i++) {
-      const start = Math.max(0, i - matchWindow);
-      const end = Math.min(i + matchWindow + 1, s2.length);
-
-      for (let j = start; j < end; j++) {
-        if (s2Matches[j] || s1[i] !== s2[j]) continue;
-        s1Matches[i] = true;
-        s2Matches[j] = true;
-        matches++;
-        break;
-      }
-    }
-
-    return { matches, s1Matches, s2Matches };
-  }
-
-  private countTranspositions(s1: string, s2: string, s1Matches: boolean[], s2Matches: boolean[]): number {
-    let transpositions = 0;
-    let k = 0;
-
-    for (let i = 0; i < s1.length; i++) {
-      if (!s1Matches[i]) continue;
-      while (!s2Matches[k]) k++;
-      if (s1[i] !== s2[k]) transpositions++;
-      k++;
-    }
-
-    return transpositions;
-  }
 }
